pyproject.toml
[project]
name = "linebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "line-bot-sdk>=3.19.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
]
